Moved qbs_for_user to better home as a static method on QB
